Home App - Update 1
Added urls for home/, frutas/, verduras, login/, carrito/ in the Home/urls.py file.

Added functions for rendering the home, frutas, verduras, login and carrito views in the Home/views.py file.

Created the templates, static, styles, images, and scripts folders for storing the html, css and image files.

Created the base.html and base.css so the rest of templates can inherit its base structure and style.

Created the .html and .css templates for the home/, frutas/, and verduras/ urls.

Added icons for the main logo, the favicon, frutas, verduras, carrito and login links, as well as some images for the home page and the list of products.