Home App - Update 2
Created a Database in MySQL with 2 tables: comuna and cliente. The schema and model can be found in the Database folder.

Updated the Django default database configuration inside the settings.py file, changing the default database from sqlite3 to MySQL.

Added the path and render function for the sign up template (registro) inside the urls.py and views.py of the Home app.

Created the front end for the login and sign up (registro) screens (.html and .css files).