Home App - Update 3
Fixed a mishandling of the database configuration in Django, generated during the last commit. Sqlite3 will continue to be the default database during development, and MySQL will be implemented for testing and production.

Deleted the Database folder in the root, since the MySQL files it contained won't be needed/used at the moment.

Applied migrations to the database db.sqlite3.

Created super user for the /admin page

Created the database models for 2 tables: comuna and cliente, inside the models.py of the Home app, for them to be migrated to the db.sqlite3 database in the root folder

Developed the views for the login and registration screens (both, backend and frontend). Still not at 100% but fairly close.

Created main menu and bank data templates (.html and .css) for the vendor profile

Created the forms.py in the Home app folder, this will be used to generate custom forms for user input and DB registration

Added "shrink-to-fit=no" to the viewport meta tag in the base.html template.

Added registration, login, main menu and bank data paths to urls.py