Home App - Update 4
Created templates for the vendor management page (.html, .css and .js).

Now it's possible to create, dele and edit vendor users in the vendor management page.

Updated the models for user and bank account

Updated the register and login views to meet the model update, now it detects whether the user is a client or a vendor and redirects them accordingly

Now it's possible to save and edit the bank account information of the vendor