Home App - Update 5
Fixed the gestionarvendedor view, now the function creates another vendor user, instead of a normal client user (is_staff attribute is set to True when the user is created). Also, the vendors list now shows only users whose is_staff attribute is set to True, instead of all the Usuario class objects in the database.

Updated the home and mainmenu views to show a welcome message to the user when he/she logs in. This also works as a hint for the user to know that the login has been successful.

Added a good-bye pop-up message when the user logs out.

Improved the pop-up alerts and messages shown to the user by importing SweetAlert2 (JavaScript library)

Tested validations on the register and login pages.

Home app seems to run fine so far. Next up is the Pedidos app.