Update 6
Started development of the Pedidos app.

Created the app directory inside the project folder, added the urls.py and forms.py files, and the templates and static folders with the necessary subfolders.

Created the Pedido model inside the models.py, and added 4 instances to the table in the DB in order to test the functionality.

Created the /pedidos view and template. It displays the shipping packages saved to the database.

Moved the /carrito URL and template from the Home app to the Pedidos app, since that's where it belongs to.