Update 7
Created a carrito.py file with a Carrito class inside to handle the shopping cart operations, which include: add items to cart, delete items from cart, clear all the items from the cart, and save them in memory while the user session is active. At the moment, only the delete items functionality is left to be implemented in the website.

frutas.html now displays actual products from the database (before it was only html).