Update 8
Fixed a couple of bugs of the shopping cart functionality. Now it's possible to add and substract products, as well as delete items from the shopping cart. This will modify the products stock field in the database. The clear cart functionality still needs some tweaking though, since it will just delete the items in the shopping cart without restoring the products stock in the database.