Update 9
Installed the Pillow library to handle the images on the website.

Did some refactoring on the frutas.html and verduras.html files, so now every product shown in those templates gets its information from the databse, including the images.

Updated the Producto model so it will store the products images. Migrated the new model to the database.

Finished the carrito template, so now it displays the products added with an image, and ordered in rows.

Updated the website buttons style on the .css files.