Update 10
Created a DetallePedido class inside the Pedidos.models, which will serve as an intermediary table to handle the many to many relationship between the Producto and Pedido tables.

Created a template and a view for the order funcionality (.html, .css, and an URL, as well as the function in views.py). It displays a summary of the items added to the shopping cart, the customer's name and address, a select widget for the payment method, the ammount of the order, and the button to place it.

When clicking on the confirmar pedido button, an order is created in the database, with the current user's id as a foreign key. However, there's an error when creating the detalle pedido register (to handle the many to many relationship). It needs fixing so the pedido register is able to link the products it contains.