Update 11
Changed the lang tags in the base.html and mainmenu.html templates from english to spanish.

Changed the Django language settings from "en-us" to "es-mx".

Fixed the error that was triggered when creating instances of the DetallePedido class in the Pedidos.views file. Now, everytime an order is placed, the function hacerpedido takes care of registering an order in the database (in the Pedido table), as well as linking the products that belong to that order, by registering them in the DetallePedido table. It also clears the shopping cart once all that is done.

Added a button to view de order details in the Pedidos module.

Added the function detallepedido inside the Pedidos.views to render the details of an order when the newly added button is pushed.

Added the function actualizarpedido inside the Pedidos.views and the function actualizarPedido inside the actualizarpedido.js file, in order to update the order status in the Pedidos module, by pressing the button Actualizar Pedido.

Added the filtrarpedidos function inside the Pedidos.views to filter the orders according to their status, by pressing the buttons at the beginning of the page (one button for each type of status) of the pedidos.html template.

Added the functions filtrarfrutas and filtrarverduras in the Pedidos.views, so now it's possible to order the products by 4 different criteria, which can be selected at the beginning of the pages frutas.html and verduras.html.

Changed the aesthetics of the select and ordenar button in the frutas.html and verduras.html templates.