Update 12
Changed the navigation bar icons to fit better with the overall theme of the website.

Added a 'go back to cart' button in the hacerpedido template.

Changed the project's timezone to America/Santiago in the settings.py. Now every model that uses an automatic DateTimeField will have the correct time to register data in the database.

Added the vendor's bank account information in the hacerpedido template, in case a customer wishes to pay via bank transfer. Also, added the estimated order delivery date in the same template.