Update 13
Modified the descripcion_producto attribute in the Producto model, instead of CharField, it's a TextField now.

Added the Comuna model inside Home.models, and inserted 35 comunas in the Comuna table. These are the comunas where the delivery service is operative.

Changed the comuna attribute in the Usuario model, instead of a CharField, now it's a ForeignKey, linked to the newly added Comuna model.

Deleted the requirement for the comuna input in the create_superuser function in the Home.models file. This allows to create a super user for the first time without entering a comuna (which would otherwise result in an error, since the comuna table wouldn't have any register yet).

Changed the Comuna option Santiago Centro to Santiago, in the Home.forms file.

Added a vendedor attribute to the CuentaBancaria model, and used it in the datosbanco function inside the Home.views, in order to show in the datosbanco.html template the last user that modified the vendor's bank information.

Added a pop up script with sweetalert 2 for the form error messages in the user register page. (registro.html).

Changed the max_lenght for the password attribute in the Usuario class, now it's 255 (which is necessary for Django to encrypt the users' passwords as long strings).

Added a verification to assess whether an user trying to access the vendor main menu and its submenus via the URL bar is staff or not, if he's not, he will be redirected to the home page.

Added a verification to assess whether the shopping cart has items or not, if it doesn't, the user won't be able to access the hacerpedido page.

Added a warning message that pops up when an authenticated user tries to access the sign up or login pages.