Update 14
Created the gastos app inside the project root folder. Added the urls and forms files inside the Gastos app folder, as well as the static and templates folders. Added the app in the project settings' installed apps section, and included the Gastos.urls.py in the project's main urls.py file.

Added the gastos function inside the Gastos.views and it's corresponding template, gastos.html (including its style gastos.css and its script gastos.js). It's not ready yet though, still needs some work for it to be fully functional.