Update 15
Added the ingresarotrosgastos function inside the Gastos.views file. Now it's possible to add aditional expenses (not related to either fruits, vegetables or bags) for a given day. These aditional expenses will be saved in a column inside the GastoProductos table.

Modified the GastoProductos and the OtroGasto classes inside the Gastos.models file, adding a couple of attributes needed to link the products expenses and the aditional expenses for a given day.