Update 17
Did some refactoring on the Gastos.forms file, so the products expenses form will only display 3 fields on the template, which are the only ones that need user input. Modified the gastos.html template accordingly.

Added icons for the filter options in the gastos.html template.

Added an input for date selection when registering a new expense for products. Before, it would add the current date automatically, giving no option to add the user's desired date.