Update 18
Added the detalleotrosgastos function inside the Gastos.views file, and the detalleotrosgastos.html inside the templates folder. Now it's possible  to view other expenses associated to a particular date from the products expenses list. This can be achieved by clicking on the Detalle Otros Gastos button in the gastos page.

Added a delete functionality for every expense in the expenses list. This includes an eliminargasto function inside the Gastos.views file, as well as an eliminarGasto function inside the gastos.js file.

Added an edit functionality for every expense in the expenses list, from the gastos.html template. This includes an editargasto function inside the Gastos.views file, as well as a new html file inside the templates folder, editargasto.html.