Update 20
Added the month filter functionality to the expenses module. This includes the filtrarmes function inside the Gastos.views file, and a path inside the Gastos.urls file.

Added the expenses filter by date range to the expenses module.

Created the Reports app, which in includes a Reportes folder, with subdirectories for the static and templates subfolders.

Added .png files for the Reports app, which were placed inside the static/images directory.

Created the reportes.html and the elegirfecha.html templates, which use these icons. These templates are used to select the type of report that the user needs.

Added a functionality to create sales reports, which can be filtered by week, month, or a custom range. This includes a path inside the Reportes.urls file, a reporteventas function inside the Reportes.views, and a reporteventas.html template.

Added a functionality to create expenses reports, which can be filtered by week, month, or a custom range. This includes a path inside the Reportes.urls file, a reportegastos function inside the Reportes.views, and a reportegastos.html template.

Added a functionality to create sales by customer reports, which can be filtered by week, month, or a custom range. This includes a path inside the Reportes.urls file, a reportecliente function inside the Reportes.views, and a reportecliente.html template.