Update 21
Fixed a GUI problem with the update order button inside the pedidos.html file (it was too long, and was overlapping into the next row on the list).

Added a go back to reports button inside the elegirfecha.html file, from the Reportes module.

Added a warning message in the detalleotrosgastos function, inside the Gastos.views file. This message pops up when the user tries to acces the other expenses details page, but there's no other expenses associated with that particular date.

Moved the add new expense button from the bottom to the top of the page, for better usability (gastos.html).

Added a view all expenses button inside the gastos.html.

Did some refactoring on the Gastos.views file, the expenses filter functions (by week, month or dates range) are now implemented in just 1 function (filtrargastos), instead of them being 3 different functions, one for each kind of filter.

Added a validation inside the reporteventas, reportecliente and reportegastos functions (Reportes.views), which alerts the user via a warning message, that the queried period has no records.

Installed the ReportLab Python library in the project's virtual environment, in order for the application to be able to export files into .pdf format.

Added the functionality to export into PDF and download into the user's PC sales and expenses reports. This includes a function inside the Reportes.views file: pdfreporte.