Update 22
Added the products module view and templates. This includes 5 functions inside the Productos.views file, as well as 3 templates (html and css), and a productos.js file.

The add product functionality is almost ready, at the moment it's not saving in the database the necessary image files when uploading them into the image field from the product form, so it will need some fixing for the next update.