Update 23
Finished the functionality to add a new product (Productos.views.agregarproducto). Now it's possible to fill in the form, add an image file and save it to the database. The app validates that it's in fact an image format, and will pop up an error message when it's not.

Added a stats section inside the products module (just the icon in the menu), which will show statistics about best sellers, etc.