Update 24
Added the django-cleanup Python library, in order to handle the deletion of unused files inside the media folder of the project. This comes at hand when the file associated with a model through a FileField or ImageField changes, i.e. an object is updated or deleted.

Added the edit product functionality, which includes the editarproducto function inside the Productos.views file, as well as its .html and .css templates. Now it's possible to update the data associated with a product, including its image.

Added the delete product functionality, this includes the eliminarproducto function inside the Productos.views file, as well as its .html and .css templates. This makes possible to delete a product from the database, including its associated image, which is stored inside the media folder.