Update 25
Fixed an error that rendered impossible to make new orders. It was caused by two functions having the same name (agregarproducto). Changed one of them to crearproducto, inside the Productos.views file.

Did some refactoring in the Pedidos.views file. Now the orders list will display orders starting from the most recent on top.

Fixed an error that produced a software failure when trying to delete products from the database, in cases where the aforementioned product is included in one or more customers' orders. Now the suprimirproducto function inside the Productos.views file, will validate wether the product is included in an order or not.

Also, changed the function's name form eliminarproducto to suprimirproducto, since there was already an eliminarproducto function inside the Pedidos.views file. This will prevent errors derived from both functions having the same name.

Modified the functions that render the fruits and vegetables lists for the user to place an order (the frutas and verduras functions, inside the Pedidos.views file). Now the app will only show products that have a stock greater than zero.