Update 26
Added the is_active attribute to the Producto class, inside Productos.models. This is used to either hide or show a product from the sales section. Also, modified the Productos.forms file and the adminproductos, crearproducto and editarproducto html templates to account for the newly added attribute.

Added the estadisticaproducto function inside the Productos.views file and its related .html and .css templates. This function uses the Chart.js javascript library in order to create charts, which in this case, show the best selling fruits and vegetables. This can be viewed by clicking on the "Estadisticas" option in the products module.