Update 27
Added the Clientes app to the project. This includes the app folder inside the project's root, and a static and templates subfolders. Also, added the newly created app to the project's installed apps, inside the settings.py, and the 'clientes/' url inside the project's urls.py file.

Added the clientes, editarcliente and eliminarcliente functions inside the Clientes.views file. This also includes the clientes and editarcliente .html and .css templates, as well as a clientes.js script to handle the deletion of customers.

Added the /clientes url in the mainmenu.html template, so the customers module can be accessed from the vendor main menu.

Added an icon and link in the base.html's navbar, so customers can access their user profile.

Added the perfilcliente function inside the Clientes.views file. This also includes a path in the Clientes.urls, a perfilcliente.html and a perfilciente.css template. This page is a submenu, which helps the user (customer) access information about his/her orders, and allows him/her to update personal information.

Added the actualizardatos, mispedidos and mipedido functions inside the Clientes.views. This includes html and css templates for each one of these functions. Actualizarpedidos enables the user (customer) to update his/her personal information, while mispedidos shows the user his/her placed orders in a list. From this list, the user can also access the details of each placed order, reaching a detailed page, which is generated by the mipedido function.