Update 28
Added a function to delete other expenses inside the Gastos.views file. This includes a path in the Gastos.urls and a JavaScript Function inside the gastos.js. This is a necessary addition to the software for it to function properly.

Did some refactoring on the gestionarvendedor .html and .css templates. Also, corrected some format errors in the detalleotrosgastos.html and the gastos.html templates.

Fixed an error in the editarvendedor.html template that made the page to render incorrectly.

Removed a line from the datosbanco.html due to it being unnecessary.

Updated the requirements.txt file.