Update 29
Cleaned up the Python code and improved the documentation by making it more easily understandable.

Did some refactoring on the Gastos.views file, and added an icon to the navigation bar inside the gastos.html template, in order to improve the software's usability.

Fixed some format issues on the index.html, registro.html and mainmenu.html templates.

Changed the chart type inside the estadisticaproducto.html file so it will be easier to comprehend.