Update 31
Fixed an error inside the Productos.views file that resulted in a crash whenever an user tried to access the products statistics page without any pre existing sales in the database. This also includes some extra lines inside the productos.html template in order to show an alert message to the user.