Update 32
Made some changes inside the hacerpedido.html template, in order to show the VAT to the user when placing an order, as well as the delivery cost. This includes calculating the tax inside the Pedidos.context_processor.py file, as well as adding the delivery cost to the total ammount.

Fixed an error that resulted in a crash when accessing via the URL bar to the hacerpedido page with no products in the shopping cart.