Update 33
Modified the charts properties inside the estadisticaproducto.html template, in order for them to display not only the quantity of products' sales, but the revenue as well. This also includes adding 2 new list variables inside the Productos.views file, in order to store the revenue data that is passed to the template for the chart construction.

Changed some minor format parameters inside the mispedidos.html, mispedidos.css templates, crearproducto.css and editarproducto.css files.